+++ BASE/HBuffer.h
@@ -26,18 +26,18 @@ public:
     virtual ~HBuffer();
 
     // These work pretty much like you'd expect.
-    virtual void write(size_t len, const HBYTE *data, bool wait = true);
+    virtual void write(uint64 len, const HBYTE *data, bool wait = true);
     void write(const HBuffer &other);
     void writeAsFormat(const HBYTE *fmt, ...);
 
-    virtual void read(size_t length, HBYTE *dest, bool wait = true);
-    virtual void read(size_t length, HBuffer &other, bool wait = true);
-    virtual void peek(size_t length, HBYTE *dest, bool wait = true);
-    virtual void peek(size_t length, HBuffer &other, bool wait = true);
-    virtual void skip(size_t length, bool wait = true);
+    virtual void read(uint64 length, HBYTE *dest, bool wait = true);
+    virtual void read(uint64 length, HBuffer &other, bool wait = true);
+    virtual void peek(uint64 length, HBYTE *dest, bool wait = true);
+    virtual void peek(uint64 length, HBuffer &other, bool wait = true);
+    virtual void skip(uint64 length, bool wait = true);
 
     // Utilities. Try to avoid checking for size.
-    size_t size() const { return data_.size(); }
+    uint64 size() const { return data_.size(); }
     bool empty() const { return size() == 0; }
     void clear() { data_.resize(0); }
     HBYTE *data() { return &data_[0]; }
@@ -45,7 +45,7 @@ public:
 protected:
     // Write max [length] bytes to the returned pointer.
     // Any other operation on this Buffer invalidates the pointer.
-    HBYTE *appendBufferSize(size_t length);
+    HBYTE *appendBufferSize(uint64 length);
 
     // TODO: Find a better internal representation, like a cord.
     std::vector<HBYTE> data_;
@@ -57,13 +57,13 @@ class HInBuffer : protected HBuffer
 {
 public:
     template <typename T>
-    void readAny(size_t length, T *dest, bool wait = true) {
+    void readAny(uint64 length, T *dest, bool wait = true) {
         if (length >= sizeof(T) && bigEndian() && sizeof(T) != 1) {
             if (length % sizeof(T) != 0)
                 throw _HException_Normal("Unaligned data size!");
 
-            size_t count = length / sizeof(T);
-            for (size_t index = 0; index < count; ++index) {
+            uint64 count = length / sizeof(T);
+            for (uint64 index = 0; index < count; ++index) {
                 read(sizeof(T), (HBYTE *)&dest[index], wait);
                 if (dest) {
                     dest[index] = swap(&dest[index]);
@@ -86,17 +86,17 @@ public:
         return temp;
     }
 
-    void skip(size_t length, bool wait = true) {
+    void skip(uint64 length, bool wait = true) {
         readAny(length, (uint8 *)nullptr, wait);
     }
 
     template <typename T>
-    void readAny(size_t length, T &dest, bool wait = true) {
+    void readAny(uint64 length, T &dest, bool wait = true) {
         read(length, dest, wait);
     }
 
 protected:
-    void checkBuffer(size_t length, bool wait = true, bool throwException = true) {
+    void checkBuffer(uint64 length, bool wait = true, bool throwException = true) {
         if (length > size()) {
             fillBuffer(length - size(), wait);
             if (throwException && wait && length > size()) {
@@ -105,7 +105,7 @@ protected:
         }
     }
 
-    virtual void fillBuffer(size_t, bool = true) {}
+    virtual void fillBuffer(uint64, bool = true) {}
 
     virtual bool bigEndian() { return false; }
 };
@@ -114,13 +114,13 @@ class HOutBuffer: protected HBuffer
 {
 public:
     template <typename T>
-    void writeAny(size_t length, T *dest, bool wait = true) {
+    void writeAny(uint64 length, T *dest, bool wait = true) {
         if (length >= sizeof(T) && bigEndian() && sizeof(T) != 1) {
             if (length % sizeof(T) != 0)
                 throw _HException_Normal("Unaligned data size!");
 
-            size_t count = length / sizeof(T);
-            for (size_t index = 0; index < count; ++index) {
+            uint64 count = length / sizeof(T);
+            for (uint64 index = 0; index < count; ++index) {
                 T temp = swap(&dest[index]);
                 write(sizeof(T), (HBYTE *)&temp, wait);
             }
@@ -139,14 +139,14 @@ public:
         writeAny(sizeof(T), &dest, wait);
     }
 
-    void pad(size_t length) {
+    void pad(uint64 length) {
         uint8 temp = 0;
-        for (size_t index = 0; index < length; ++index) {
+        for (uint64 index = 0; index < length; ++index) {
             writeOne(&temp);
         }
     }
 
-    virtual void flushBuffer(size_t = 0, bool = true) {}
+    virtual void flushBuffer(uint64 = 0, bool = true) {}
 
 protected:
     virtual bool bigEndian() { return false; }
