+++ IO/FDOutBuffer.h
@@ -13,10 +13,10 @@ namespace IO
 
         void setTimeout(int timeoutms);
 
-        void flushBuffer(size_t length = 0, bool wait = true) override;
+        void flushBuffer(uint64 length = 0, bool wait = true) override;
 
     protected:
-        void write(size_t length, const HBYTE *data, bool wait = true) override;
+        void write(uint64 length, const HBYTE *data, bool wait = true) override;
         bool bigEndian() override { return true; }
 
     private:
